Add API for getting endianness of a binary resource bundle
